+++ crates/http/src/request.rs
@@ -468,6 +468,7 @@ impl ParseHttp for Server {
                                     lang.split_once(';').map_or(lang, |(l, _)| l)
                                 })
                                 .unwrap_or("en"),
+                            session.remote_ip,
                         )
                         .await
                         .map(|response| {